Updated helper method structures
