+++ src/services/data/TweetService.ts
@@ -15,10 +15,11 @@ import RawRetweeters from '../../types/raw/tweet/Retweeters';
 import * as Urls from '../helper/Urls';
 
 // EXTRACTORS
-import * as Extractors from "../helper/Extractors";
+import * as TweetExtractors from "../helper/extractors/Tweets";
 
 // DESERIALIZERS
-import * as Deserializers from '../helper/Deserializers';
+import * as UserDeserializers from '../helper/deserializers/Users';
+import * as TweetDeserializers from '../helper/deserializers/Tweets';
 
 // PARSERS
 import { toQueryString } from '../helper/Parser';
@@ -48,13 +49,13 @@ export class TweetService extends FetcherService {
         let res = await this.request<RawTweets>(Urls.tweetsUrl(toQueryString(filter), count, cursor), false).then(res => res.data);
 
         // Extracting data
-        let data = Extractors.extractTweets(res);
+        let data = TweetExtractors.extractTweets(res);
 
         // Caching data
         this.cacheData(data);
 
         // Parsing data
-        let tweets = data.required.map(item => Deserializers.toTweet(item));
+        let tweets = data.required.map(item => TweetDeserializers.toTweet(item));
 
         return {
             list: tweets,
@@ -79,13 +80,13 @@ export class TweetService extends FetcherService {
         let res = await this.request<RawTweet>(Urls.tweetDetailsUrl(tweetId), false).then(res => res.data);
 
         // Extracting data
-        let data = Extractors.extractTweet(res, tweetId);
+        let data = TweetExtractors.extractTweet(res, tweetId);
 
         // Caching data
         this.cacheData(data);
 
         // Parsing data
-        let tweet = Deserializers.toTweet(data.required[0]);
+        let tweet = TweetDeserializers.toTweet(data.required[0]);
 
         return tweet;
     }
@@ -111,13 +112,13 @@ export class TweetService extends FetcherService {
         let res = await this.request<RawLikers>(Urls.tweetLikesUrl(tweetId, count, cursor)).then(res => res.data);
 
         // Extracting data
-        let data = Extractors.extractTweetLikers(res);
+        let data = TweetExtractors.extractTweetLikers(res);
 
         // Caching data
         this.cacheData(data);
 
         // Parsing data
-        let users = data.required.map(item => Deserializers.toUser(item));
+        let users = data.required.map(item => UserDeserializers.toUser(item));
 
         return {
             list: users,
@@ -146,13 +147,13 @@ export class TweetService extends FetcherService {
         let res = await this.request<RawRetweeters>(Urls.tweetRetweetUrl(tweetId, count, cursor)).then(res => res.data);
 
         // Extracting data
-        let data = Extractors.extractTweetRetweeters(res);
+        let data = TweetExtractors.extractTweetRetweeters(res);
 
         // Caching data
         this.cacheData(data);
 
         // Parsing data
-        let users = data.required.map(item => Deserializers.toUser(item));
+        let users = data.required.map(item => UserDeserializers.toUser(item));
 
         return {
             list: users,
@@ -175,13 +176,13 @@ export class TweetService extends FetcherService {
         let res = await this.request<RawTweet>(Urls.tweetRepliesUrl(tweetId, cursor)).then(res => res.data);
 
         // Extracting data
-        let data = Extractors.extractTweetReplies(res, tweetId);
+        let data = TweetExtractors.extractTweetReplies(res, tweetId);
 
         // Caching data
         this.cacheData(data);
 
         // Parsing data
-        let tweets = data.required.map(item => Deserializers.toTweet(item));
+        let tweets = data.required.map(item => TweetDeserializers.toTweet(item));
 
         return {
             list: tweets,
